styles: js file path
wip
